requirements.txt
python_dotenv>=1.1.0, <2.0.0
aiohttp>=3.12.4, <4.0.0
azure-identity>=1.23.0, <2.0.0
azure-ai-projects==1.0.0b11
azure-ai-agents==1.0.0
pymongo
azure-ai-inference